+++ src/Tile.ts
@@ -5,8 +5,16 @@ export type Tag = string | Node
 
 export class Tile {
 	node: Node
+	_onEnter: (tile: Tile) => any
+	_onExit: (tile: Tile) => Promise<any>
+	_onCancelExit: (tile: Tile) => any
+	_removingChildren: Set<Tile>
 
 	constructor (tag: Tag, isText = false) {
+		this._onEnter = () => {}
+		this._onExit = () => Promise.resolve()
+		this._onCancelExit = () => {}
+		this._removingChildren = new Set([])
 		if (typeof tag !== 'string') {
 			this.node = tag
 			return
@@ -86,11 +94,18 @@ export class Tile {
 		}
 		return tile
 	}
-	children (children: Tile[] = []) {
+	children (children: Tile[] = [], childrenToRemove: Tile[] = []) {
 		const node = this.node
+		childrenToRemove.forEach(childToRemove => this._removingChildren.add(childToRemove))
+		children.filter(child => !childrenToRemove.includes(child)).forEach(child => {
+			if (!this._removingChildren.has(child)) return
+			this._removingChildren.delete(child)
+			child._onCancelExit(this)
+		})
 		if (node instanceof Text) return this
 		let a = Array.prototype.slice.call(node.childNodes)
 		const b = children.map(tile => tile.node)
+		const childrenByNode = new Map(children.map(child => [child.node, child]))
 		if (a.length + b.length === 0) return this
 		a.filter(element => !b.includes(element)).forEach(element => node.removeChild(element))
 		a = Array.prototype.slice.call(node.childNodes)
@@ -101,12 +116,18 @@ export class Tile {
 		}
 		b_c = b_i
 		while (b_i++ < b.length - 1) {
+			const ch = childrenByNode.get(b[b_i])
+			if (childrenToRemove.includes(ch)) continue
 			node.appendChild(b[b_i])
+			ch._onEnter(ch)
 		}
 
 		for (let b_i = b_c - 1; b_i >= 0; b_i--) {
 			if (!a.includes(b[b_i])) {
+				const ch = childrenByNode.get(b[b_i])
+				if (childrenToRemove.includes(ch)) continue
 				node.insertBefore(b[b_i], b[b_i + 1])
+				ch._onEnter(ch)
 				continue
 			}
 			if (a.indexOf(b[b_i]) !== a.indexOf(b[b_c]) - 1) {
@@ -115,6 +136,16 @@ export class Tile {
 
 			b_c = b_i
 		}
+		childrenToRemove.forEach(async (childToRemove) => {
+			if (!Array.from(this.node.childNodes).includes(childToRemove.node)) return
+			const onExit = childToRemove._onExit(childToRemove)
+			if (!(onExit instanceof Promise)) throw new Error('onExit must return a promise')
+			await onExit
+			if (!Array.from(this.node.childNodes).includes(childToRemove.node)) return
+			if (!this._removingChildren.has(childToRemove)) return
+			this._removingChildren.delete(childToRemove)
+			node.removeChild(childToRemove.node)
+		})
 		return this
 	}
 	adopt (child: Tag) {
@@ -149,6 +180,21 @@ export class Tile {
 	}
 	remove () {
 		this.node.parentElement.removeChild(this.node)
+	}
+	onEnter (_onEnter) {
+		if (!(_onEnter instanceof Function)) throw new Error('onEnter must receive a function')
+		this._onEnter = _onEnter.bind(this)
+		return this
+	}
+	onExit (_onExit) {
+		if (!(_onExit instanceof Function)) throw new Error('onExit must receive a function')
+		this._onExit = _onExit.bind(this)
+		return this
+	}
+	onCancelExit (_onCancelExit) {
+		if (!(_onCancelExit instanceof Function)) throw new Error('onCancelExit must receive a function')
+		this._onCancelExit = _onCancelExit.bind(this)
+		return this
 	}
 }